+++ components/makeMyManual/makeMyManual.tsx
@@ -5,32 +5,25 @@ import React from 'react'
 import { useQuery } from '@tanstack/react-query'
 import classNames from 'classnames/bind'
 import { useForm } from 'react-hook-form'
-import { useRecoilState, useRecoilValue, useResetRecoilState } from 'recoil'
+import { useRecoilValue } from 'recoil'
 
 import CharacterBox from '@/components/characterBox/characterBox'
 import CommonBtn from '@/components/commonBtn/commonBtn'
-import { Input } from '@/components/commonInput/input'
 import FormBox from '@/components/formBox/formBox'
-import Inventory from '@/components/inventory/inventory'
-import ResetBtn from '@/components/resetBtn/resetBtn'
 
 import useBreakpoint from '@/hooks/useBreakpoint.hooks'
 import { useFunnel } from '@/hooks/useFunnel'
 
 import { ButtonStyle } from '../commonBtn/commonBtn.types'
-import KeywordBtnList from '../keywordBtnList/keywordBtnList'
-import MyDescriptionCardList from '../myDescriptionCardList/myDescriptionCardList'
-import RadarChartContainer from '../radarChart/radarChartContainer'
 import { fetchMyProfileRegisterData } from './makeMyManual.api'
 import styles from './makeMyManual.module.scss'
+import MakeMyManualFunnel from './makeMyManualFunnel/makeMyManualFunnel'
 import {
   selectedBodyItemState,
   selectedExpressionItemState,
   selectedFaceItemState,
-  selectedKeywordsState,
   selectedSetState,
 } from './store/makeMyManual.atom'
-import { originKeywordPercentsSelector } from './store/originKeywordPercents.selecter'
 
 const cx = classNames.bind(styles)
 
@@ -42,8 +35,8 @@ export default function MakeMyManual() {
   })
 
   const {
-    register,
     handleSubmit,
+    register,
     formState: { isSubmitting },
   } = useForm()
   const { Funnel, step, goPrev, goNext } = useFunnel(
@@ -54,22 +47,8 @@ export default function MakeMyManual() {
   const selectedBodyItem = useRecoilValue(selectedBodyItemState)
   const selectedExpressionItem = useRecoilValue(selectedExpressionItemState)
   const selectedSet = useRecoilValue(selectedSetState)
-  const [selectedKeywords, setSelectedKeywords] = useRecoilState(
-    selectedKeywordsState,
-  )
-  const originKeywordPercents = useRecoilValue(originKeywordPercentsSelector)
 
   const isTablet: boolean = useBreakpoint({ query: '(max-width: 1199px)' })
-  const isMobile: boolean = useBreakpoint({ query: '(max-width: 768px)' })
-  let inputVariant: string
-
-  if (isTablet) {
-    inputVariant = 'medium'
-  } else if (isMobile) {
-    inputVariant = 'small'
-  } else {
-    inputVariant = 'large'
-  }
 
   // const validationRules = {
   //   required: '필수 입력입니다.',
@@ -86,23 +65,6 @@ export default function MakeMyManual() {
         body: selectedBodyItem,
         expression: selectedExpressionItem,
       }
-
-  const rectangleLayout = {
-    frameSize: 350,
-    radarSize: 200,
-  }
-
-  const resetFace = useResetRecoilState(selectedFaceItemState)
-  const resetBody = useResetRecoilState(selectedBodyItemState)
-  const resetExpression = useResetRecoilState(selectedExpressionItemState)
-  const resetSet = useResetRecoilState(selectedSetState)
-
-  const handleResetBtnClick = () => {
-    resetFace()
-    resetBody()
-    resetExpression()
-    resetSet()
-  }
 
   const onClickSubmit = () => {
     goNext()
@@ -126,59 +88,12 @@ export default function MakeMyManual() {
                 selectedItems={selectedItems}
               />
             )}
-
-            <Funnel>
-              <Funnel.Step name="nickname">
-                <div className={cx('input')}>
-                  <Input variant={inputVariant}>
-                    <Input.TextField
-                      id="nickname"
-                      register={register('nickname')}
-                    />
-                  </Input>
-                </div>
-              </Funnel.Step>
-              <Funnel.Step name="character">
-                <div className={cx('inventory')}>
-                  <Inventory
-                    resetBtn={<ResetBtn onClick={handleResetBtnClick} />}
-                  />
-                </div>
-              </Funnel.Step>
-              <Funnel.Step name="manual">
-                <div className={cx('manualWrap')}>
-                  <MyDescriptionCardList register={register('manual')} />
-                </div>
-              </Funnel.Step>
-
-              <Funnel.Step name="keyword">
-                <div className={cx('keywords')}>
-                  <KeywordBtnList
-                    selectedKeywords={selectedKeywords}
-                    setSelectedKeywords={setSelectedKeywords}
-                    // keywords={data?.itemsData?.exampleKeywords}
-                  />
-                </div>
-              </Funnel.Step>
-
-              <Funnel.Step name="statGraph">
-                {originKeywordPercents && (
-                  <RadarChartContainer
-                    radarType="NJNS"
-                    originKeywordPercents={originKeywordPercents}
-                    otherKeywordPercents={{}}
-                    frameSize={rectangleLayout.frameSize}
-                    radarSize={rectangleLayout.radarSize}
-                    framePadding={
-                      rectangleLayout.frameSize - rectangleLayout.radarSize
-                    }
-                    hasOthers={false}
-                  />
-                )}
-              </Funnel.Step>
-            </Funnel>
+            <MakeMyManualFunnel
+              Funnel={Funnel}
+              register={register}
+              isInvalid={false}
+            />
           </div>
-
           <div className={cx('btn')}>
             <CommonBtn
               type="submit"
